Photo Collage Assignment.
Photo collage assignment. I know this is a simlpe concept but for some reason I could not get past the lerning curve. This was really challenging and after looking at online tutorials and the code we did in class, I feel like I could use a lot of help with this if possible.
-Natalia